Promote portability with int32_t

+++ deps/calc.h
@@ -2,6 +2,6 @@
 #define CALC_H_INCLUDED
 #include "stack.h"
 #include "edutils.h"
-int calculate(char *input, long *result, void (*message_callback)(const char const *message));
+int32_t calculate(char *input, int32_t *result, void (*message_callback)(const char const *message));
 
 #endif
